prueba flyway

+++ src/main/java/co/edu/unimagdalena/busreserve/domine/entities/User.java
@@ -3,7 +3,6 @@ package co.edu.unimagdalena.busreserve.domine.entities;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.Id;
 
 import java.time.LocalDateTime;
 
@@ -18,7 +17,7 @@ import java.time.LocalDateTime;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @Column()
     private String name;
     @Column()
